WebScraping/requirements.txt
python-decouple
beautifulsoup4
lxml